% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_conditional_power.R
\name{get_conditional_power}
\alias{get_conditional_power}
\title{Get the conditional power based on the interim result}
\usage{
get_conditional_power(zt, cb, x2)
}
\arguments{
\item{zt:}{test statistic at the first stage}

\item{cb:}{critical boundary at the second stage}

\item{x2:}{information vector}
}
\value{
conditional power
}
\description{
This function computes the conditional power
based on the interim results
}
