% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_promising_zone.R
\name{get_promising_zone}
\alias{get_promising_zone}
\title{Get the promising zone and sample size adaptation}
\usage{
get_promising_zone(beta, cb, x2, s.size, s.max)
}
\arguments{
\item{beta:}{type II error}

\item{cb:}{critical boundary at the second stage}

\item{x2:}{information vector}

\item{s.size:}{sample size designed at the beginning}

\item{s.max:}{maximum sample size based on the budget}
}
\value{
a list of promising zone and a table for sample size adaptation based on interim test statistics
}
\description{
This function computes the promising zone and
provide newly adapted sample size for second stage
if the interim result is located in the promising zone
}
