% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_two_stage_boundary.R
\name{get_two_stage_boundary}
\alias{get_two_stage_boundary}
\title{Generate the critical boundaries of two-stage design}
\usage{
get_two_stage_boundary(x1, x2, x3)
}
\arguments{
\item{x1:}{type 1 error (alpha)}

\item{x2:}{information vector}

\item{x3:}{alpha spending function. Pb and OBFb}
}
\value{
critical boundaries of two-stage design
}
\description{
This function creates the critical boundaries of two-stage design
based on the alpha spending function
user can specify the alpha spending function by themself
}
